pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "poe2-opportunity-monitor"
version = "0.1.0"
description = "Monitor POE2 channels for profitable opportunities and alert via Telegram."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "openai>=1.51,<2.0",
  "python-dotenv>=1.0,<2.0",
  "python-telegram-bot>=21.0,<22.0",
  "pydantic>=2.8,<3.0",
  "uvicorn>=0.30,<1.0",
]

[project.scripts]
poe2-monitor = "poe2_monitor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
